Read the pid of the active HAproxy process immediately before needing it, instead of polling afterwards.
The impetus for the change is that warnings in the config file pass validation, but register as an error when starting the process. In that case the polling code will never be executed and we'll be out of sync. By reading the pid before we need it, we're more likely to have the right pid.

Note: in both cases we can still have orphaned haproxy processes if the user is calling reload very fast in succession. The solution for that would be to read the PID from the spawned process, and track lifetime of the process better,

var process = run( ...., function(){
  // this is the pid of the created process.
   process.pid
}

process.on('exit ...